+++ src/Uno.UI.RuntimeTests/Tests/Windows_Storage/Given_StorageFolder_Native.windows.cs
@@ -1,5 +1,4 @@
 ﻿#if WINDOWS_UWP
-
 #nullable enable
 
 using System;
@@ -12,7 +11,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_Storage
 	[TestClass]
 	public class Given_StorageFolder_Native : Given_StorageFolder_Native_Base
 	{
-		private StorageFolder _rootFolder;
+		private StorageFolder? _rootFolder;
 
 		protected override async Task<StorageFolder> GetRootFolderAsync()
 		{
@@ -23,7 +22,10 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_Storage
 
 		protected override async Task CleanupRootFolderAsync()
 		{
-			await _rootFolder.DeleteAsync();
+			if (_rootFolder != null)
+			{
+				await _rootFolder.DeleteAsync();
+			}
 		}
 	}
 }
